+++ CHANGELOG.md
@@ -18,6 +18,7 @@ For more information about this file see also [Keep a Changelog](http://keepacha
   - You can compile all pkgdown pages locally at any time with `make pkgdocs`.
 - The ERA5 NC extraction function can now handle multi-site instead of one.
 - Added function for merging images from the same tiling system (MODIS, GLANCE, ICESat-2, HLS, etc.).
+- Added function for converting images towards the GDAL-supported formats (H5, NetCDF, HDF4, GeoTIFF, etc .).
 - Added `AmeriFlux_met_process()` function with ERA5 fallback for AmeriFlux meteorological data processing and ensemble generation
 
 ### Fixed